+++ tests/tests.py
@@ -13,6 +13,7 @@ import hashlib
 import urllib
 from shutil import rmtree, which
 from shlex import quote
+import pytest
 
 from snakemake import snakemake
 from snakemake.shell import shell
@@ -35,6 +36,7 @@ def md5sum(filename):
     return hashlib.md5(data).hexdigest()
 
 
+# test skipping
 def is_connected():
     try:
         urllib.request.urlopen("http://www.google.com", timeout=1)
@@ -42,6 +44,28 @@ def is_connected():
     except urllib.request.URLError:
         return False
 
+
+def is_ci():
+    return "CI" in os.environ
+
+
+def has_gcloud_service_key():
+    return "GCLOUD_SERVICE_KEY" in os.environ
+
+
+def has_gcloud_cluster():
+    return "GCLOUD_CLUSTER" in os.environ
+
+gcloud = pytest.mark.skipif(not is_connected()
+                                 or not has_gcloud_service_key()
+                                 or not has_gcloud_cluster(),
+                                 reason="Skipping GCLOUD tests because not on "
+                                        "CI, no inet connection or not logged "
+                                        "in to gcloud.")
+
+connected = pytest.mark.skipif(not is_connected(), reason="no internet connection")
+
+ci = pytest.mark.skipif(not is_ci(), reason="not in CI")
 
 def copy(src, dst):
     if os.path.isdir(src):
@@ -52,21 +76,15 @@ def copy(src, dst):
 
 def run(path,
         shouldfail=False,
-        needs_connection=False,
         snakefile="Snakefile",
         subpath=None,
+        no_tmpdir=False,
         check_md5=True, cores=3, **params):
     """
     Test the Snakefile in path.
     There must be a Snakefile in the path and a subdirectory named
     expected-results.
     """
-
-    if needs_connection and not is_connected():
-        print("Skipping test because of missing internet connection",
-              file=sys.stderr)
-        return False
-
     results_dir = join(path, 'expected-results')
     snakefile = join(path, snakefile)
     assert os.path.exists(snakefile)
@@ -95,7 +113,7 @@ def run(path,
         # run snakemake
         success = snakemake(snakefile,
                             cores=cores,
-                            workdir=tmpdir,
+                            workdir=path if no_tmpdir else tmpdir,
                             stats="stats.txt",
                             config=config, **params)
         if shouldfail:
@@ -181,6 +199,12 @@ def test14():
 def test15():
     run(dpath("test15"))
 
+def test_directory():
+    run(dpath("test_directory"), targets=['downstream', 'symlinked_input', "child_to_input", "some/dir-child", "some/shadow"])
+    run(dpath("test_directory"), targets=['file_expecting_dir'], shouldfail = True)
+    run(dpath("test_directory"), targets=['dir_expecting_file'], shouldfail = True)
+    run(dpath("test_directory"), targets=['child_to_other'], shouldfail = True)
+
 def test_ancient():
     run(dpath("test_ancient"), targets=['D', 'old_file'])
 
@@ -254,8 +278,9 @@ def test_persistent_dict():
         pass
 
 
+@connected
 def test_url_include():
-    run(dpath("test_url_include"), needs_connection=True)
+    run(dpath("test_url_include"))
 
 
 def test_touch():
@@ -341,6 +366,9 @@ def test_script():
 def test_shadow():
     run(dpath("test_shadow"))
 
+def test_shadow_prefix():
+    run(dpath("test_shadow_prefix"), shadow_prefix = "shadowdir")
+    run(dpath("test_shadow_prefix"), shadow_prefix = "shadowdir", cluster="./qsub")
 
 def test_until():
     run(dpath("test_until"),
@@ -444,6 +472,7 @@ def test_spaces_in_fnames():
 #         pass
 
 
+@connected
 def test_remote_ncbi_simple():
     try:
         import Bio
@@ -454,6 +483,7 @@ def test_remote_ncbi_simple():
     except ImportError:
         pass
 
+@connected
 def test_remote_ncbi():
     try:
         import Bio
@@ -465,6 +495,7 @@ def test_remote_ncbi():
         pass
 
 
+@ci
 def test_remote_irods():
     if os.environ.get("CI") == "true":
         run(dpath("test_remote_irods"))
@@ -561,6 +592,7 @@ def test_issue260():
 
 def test_default_remote():
      run(dpath("test_default_remote"),
+         cores=1,
          default_remote_provider="S3Mocked",
          default_remote_prefix="test-remote-bucket")
 
@@ -569,6 +601,8 @@ def test_run_namedlist():
     run(dpath("test_run_namedlist"))
 
 
+@connected
+@ci
 def test_remote_gs():
     if not "CI" in os.environ:
         run(dpath("test_remote_gs"))
@@ -576,13 +610,17 @@ def test_remote_gs():
         print("skipping test_remove_gs in CI")
 
 
+@connected
 def test_remote_log():
     run(dpath("test_remote_log"), shouldfail=True)
 
 
+@connected
 def test_remote_http():
     run(dpath("test_remote_http"))
 
+
+@connected
 def test_remote_http_cluster():
     run(dpath("test_remote_http"), cluster=os.path.abspath(dpath("test14/qsub")))
 
@@ -590,8 +628,13 @@ def test_profile():
     run(dpath("test_profile"))
 
 
+@connected
 def test_singularity():
     run(dpath("test_singularity"), use_singularity=True)
+
+@connected
+def test_singularity_conda():
+    run(dpath("test_singularity_conda"), use_singularity=True, use_conda=True)
 
 
 def test_issue612():
@@ -614,27 +657,76 @@ def test_log_input():
     run(dpath("test_log_input"))
 
 
-def test_gcloud():
-    if "CI" in os.environ and "GCLOUD_SERVICE_KEY" in os.environ:
-        cluster = os.environ["GCLOUD_CLUSTER"]
-        try:
+@pytest.fixture(scope="module")
+def gcloud_cluster():
+    class Cluster:
+        def __init__(self):
+            self.cluster = os.environ["GCLOUD_CLUSTER"]
+            self.bucket_name = 'snakemake-testing-{}'.format(self.cluster)
+
             shell("""
-            sudo $GCLOUD container clusters create {cluster} --num-nodes 3 --scopes storage-rw --zone us-central1-a --machine-type f1-micro
-            sudo $GCLOUD container clusters get-credentials {cluster} --zone us-central1-a
+            $GCLOUD container clusters create {self.cluster} --num-nodes 3 --scopes storage-rw --zone us-central1-a --machine-type f1-micro
+            $GCLOUD container clusters get-credentials {self.cluster} --zone us-central1-a
+            $GSUTIL mb gs://{self.bucket_name}
             """)
-            run(dpath("test_kubernetes"))
-            run(dpath("test_kubernetes"), use_conda=True)
-            run(dpath("test_kubernetes"), use_singularity=True)
-            run(dpath("test_kubernetes"), use_singularity=True, use_conda=True)
-        finally:
-            shell("sudo $GCLOUD container clusters delete {cluster} --zone us-central1-a --quiet")
-    print("Skipping google cloud test")
+
+        def __del__(self):
+            shell("""
+            $GCLOUD container clusters delete {self.cluster} --zone us-central1-a --quiet || true
+            $GSUTIL rm -r gs://{self.bucket_name} || true
+            """)
+
+        def run(self, **kwargs):
+            try:
+                run(dpath("test_kubernetes"),
+                    kubernetes="default",
+                    default_remote_provider="GS",
+                    default_remote_prefix=self.bucket_name,
+                    no_tmpdir=True,
+                    **kwargs)
+            except Exception as e:
+                shell("for p in `kubectl get pods | grep ^snakejob- | cut -f 1 -d ' '`; do kubectl logs $p; done")
+                raise e
+
+        def reset(self):
+            shell('$GSUTIL rm -r gs://{self.bucket_name}/* || true')
+    return Cluster()
 
 
+
+@gcloud
+def test_gcloud_plain(gcloud_cluster):
+    gcloud_cluster.reset()
+    gcloud_cluster.run()
+
+
+@gcloud
+@pytest.mark.skip(reason="need a faster cloud compute instance to run this")
+def test_gcloud_conda(gcloud_cluster):
+    gcloud_cluster.reset()
+    gcloud_cluster.run(use_conda=True)
+
+
+@gcloud
+@pytest.mark.skip(reason="need a faster cloud compute instance to run this")
+def test_gcloud_singularity(gcloud_cluster):
+    gcloud_cluster.reset()
+    gcloud_cluster.run(use_singularity=True)
+
+
+@gcloud
+@pytest.mark.skip(reason="need a faster cloud compute instance to run this")
+def test_gcloud_conda_singularity(gcloud_cluster):
+    gcloud_cluster.reset()
+    gcloud_cluster.run(use_singularity=True, use_conda=True)
+
+
+@connected
 def test_cwl():
     run(dpath("test_cwl"))
 
 
+@connected
 def test_cwl_singularity():
     run(dpath("test_cwl"), use_singularity=True)
 
@@ -680,6 +772,27 @@ def test_issue850():
 def test_issue860():
     run(dpath("test_issue860"), cluster="./qsub", targets=["done"])
 
+
+def test_issue894():
+    run(dpath("test_issue894"))
+
+def test_issue584():
+    run(dpath("test_issue584"))
+
+def test_issue912():
+    run(dpath("test_issue912"))
+
+def test_job_properties():
+    run(dpath("test_job_properties"), cluster="./qsub.py")
+
+def test_issue916():
+    run(dpath("test_issue916"))
+
+def test_issue930():
+    run(dpath("test_issue930"), cluster="./qsub")
+
+def test_issue635():
+    run(dpath("test_issue635"), use_conda=True, check_md5=False)
 
 if __name__ == '__main__':
     import nose